pyproject.toml
[tool.poetry]
name = "sampling"
version = "0.1.0"
description = ""
authors = ["artyom <artyomevseev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.0.0"
python-dotenv = "^1.0.1"
psycopg = "^3.1.18"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
